+++ src/gfx/font_rendering/shape_drawing.hpp
@@ -3,6 +3,7 @@
 #include "./glyph_info.hpp"
 #include "./shape_pen.hpp"
 #include "./texture_atlas.hpp"
+#include "editor/highlight.hpp"
 #include "glm/ext/vector_float2.hpp"
 #include <string>
 #include <mdspan>
@@ -16,11 +17,16 @@ struct ShapeDrawing {
   using GlyphInfoMap = std::unordered_map<char32_t, GlyphInfo>;
   GlyphInfoMap glyphInfoMap;
 
+  using UnderlineGlyphInfoMap = std::unordered_map<UnderlineType, GlyphInfo>;
+  UnderlineGlyphInfoMap underlineGlyphInfoMap;
+
   ShapeDrawing() = default;
   ShapeDrawing(glm::vec2 charSize, float dpiScale);
 
   // returns nullptr if not implemented
-  const GlyphInfo* GetGlyphInfo(const std::string& text, TextureAtlas<false>& textureAtlas);
+  const GlyphInfo*
+  GetGlyphInfo(const std::string& text, TextureAtlas<false>& textureAtlas);
 
-  // std::chrono::nanoseconds totalTime = std::chrono::nanoseconds(0);
+  const GlyphInfo*
+  GetGlyphInfo(UnderlineType underlineType, TextureAtlas<false>& textureAtlas);
 };